upload logic and async proxy fallback
